server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.64.0"
protobuf = "^5.27.0"
pyttsx3 = "^2.90"
pydub = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
